@@ -1,11 +1,11 @@
-//go:build !windows
-// +build !windows
+//go:build !windows && cgo
+// +build !windows,cgo
 
 package keystore
 
 import (
+	"github.com/chernyshev/gotron-sdk/pkg/address"
 	"github.com/ethereum/go-ethereum/crypto/secp256k1"
-	"github.com/fbsobreira/gotron-sdk/pkg/address"
 )
 
 func RecoverPubkey(hash []byte, signature []byte) (address.Address, error) {
