@@ -7,16 +7,16 @@ import (
 	"os"
 
 	"github.com/btcsuite/btcd/btcec/v2"
+	"github.com/chernyshev/gotron-sdk/pkg/account"
+	"github.com/chernyshev/gotron-sdk/pkg/address"
+	"github.com/chernyshev/gotron-sdk/pkg/common"
+	c "github.com/chernyshev/gotron-sdk/pkg/common"
 	"github.com/fatih/color"
-	"github.com/fbsobreira/gotron-sdk/pkg/account"
-	"github.com/fbsobreira/gotron-sdk/pkg/address"
-	"github.com/fbsobreira/gotron-sdk/pkg/common"
-	c "github.com/fbsobreira/gotron-sdk/pkg/common"
 	"golang.org/x/crypto/ssh/terminal"
 
-	"github.com/fbsobreira/gotron-sdk/pkg/ledger"
-	"github.com/fbsobreira/gotron-sdk/pkg/mnemonic"
-	"github.com/fbsobreira/gotron-sdk/pkg/store"
+	"github.com/chernyshev/gotron-sdk/pkg/ledger"
+	"github.com/chernyshev/gotron-sdk/pkg/mnemonic"
+	"github.com/chernyshev/gotron-sdk/pkg/store"
 	"github.com/spf13/cobra"
 	"github.com/tyler-smith/go-bip39"
 )
