@@ -12,12 +12,12 @@ import (
 	"strings"
 	"time"
 
+	"github.com/chernyshev/gotron-sdk/pkg/client"
+	"github.com/chernyshev/gotron-sdk/pkg/client/transaction"
+	"github.com/chernyshev/gotron-sdk/pkg/common"
+	c "github.com/chernyshev/gotron-sdk/pkg/common"
+	"github.com/chernyshev/gotron-sdk/pkg/store"
 	color "github.com/fatih/color"
-	"github.com/fbsobreira/gotron-sdk/pkg/client"
-	"github.com/fbsobreira/gotron-sdk/pkg/client/transaction"
-	"github.com/fbsobreira/gotron-sdk/pkg/common"
-	c "github.com/fbsobreira/gotron-sdk/pkg/common"
-	"github.com/fbsobreira/gotron-sdk/pkg/store"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 	"github.com/spf13/cobra/doc"
@@ -150,8 +150,8 @@ func init() {
 var (
 	// VersionWrapDump meant to be set from main.go
 	VersionWrapDump = ""
-	versionLink     = "https://api.github.com/repos/fbsobreira/gotron-sdk/releases/latest"
-	versionTagLink  = "https://api.github.com/repos/fbsobreira/gotron-sdk/git/ref/tags/"
+	versionLink     = "https://api.github.com/repos/chernyshev/gotron-sdk/releases/latest"
+	versionTagLink  = "https://api.github.com/repos/chernyshev/gotron-sdk/git/ref/tags/"
 	versionFormat   = regexp.MustCompile("v[0-9]+-[a-z0-9]{7}")
 )
 
