@@ -7,7 +7,7 @@
 package api
 
 import (
-	core "github.com/fbsobreira/gotron-sdk/pkg/proto/core"
+	core "github.com/chernyshev/gotron-sdk/pkg/proto/core"
 	_ "google.golang.org/genproto/googleapis/api/annotations"
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
@@ -1688,7 +1688,7 @@ func (x *TimePaginatedMessage) GetLimit() int64 {
 	return 0
 }
 
-//deprecated
+// deprecated
 type AccountNetMessage struct {
 	state         protoimpl.MessageState
 	sizeCache     protoimpl.SizeCache
