@@ -7,7 +7,7 @@
 package api
 
 import (
-	core "github.com/fbsobreira/gotron-sdk/pkg/proto/core"
+	core "github.com/chernyshev/gotron-sdk/pkg/proto/core"
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
